Refactored a bunch of rendering for dynamic content. Add support for custom components on index page.
