+++ src/App.jsx
@@ -1,72 +1,32 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import {
   BrowserRouter as Router,
   Switch,
-  Route,
-  Link
+  Route
 } from "react-router-dom";
 import Helmet from 'react-helmet';
 
 import resolveAssetUrl from './utils/assetUrlResolver';
-import fetchEntries from './utils/dataEntries';
 import './App.scss';
 import {
   Footer,
-  RadioPlayer,
-  VideoChatHider,
-  ShowList,
   Header,
   SubPage,
   IndexPage,
-  Sponsors
 } from './components';
-import { INDEX_PAGE, CONTENT_PAGE, NAVIGATION } from './constants/contentTypes';
-// import useLiveShowListId from './utils/liveShows';
-// import useShowList from './utils/shows';
 import { pageview } from './utils/analytics';
-import contentful from './utils/contentful';
 
 export default () => {
-  const content = fetchEntries({
-    content_type: CONTENT_PAGE,
-  }).data;
-
-  const nav = fetchEntries({
-    content_type: NAVIGATION,
-  }).data;
-
-  // const liveShowListId = useLiveShowListId();
-  // const showList = useShowList(liveShowListId);
-
-  // TODO: create a hook for fetching index content.
-  const [indexContent, setIndexContent] = useState(null);
-
-  useEffect(() => {
-    const fetchIndex = () => contentful.getEntries({ content_type: INDEX_PAGE })
-      .then(res => setIndexContent(res.items[0].fields.content));
-    fetchIndex();
-  }, []);
-
   pageview();
 
   return (
     <div className="App">
       <Router>
-        <Header>
-          <ul>
-            <li>
-              <Link to="/">Etusivu</Link>
-            </li>
-            {nav && nav.items && nav.items[0].fields.pages.map((item, idx) =>
-              <li key={idx}>
-                <Link to={`/${item.fields.slug.toLowerCase()}`}>{item.fields.name}</Link>
-              </li>)}
-          </ul>
-        </Header>
+        <Header />
         <div className="Container">
           <Switch>
             <Route path="/:id">
-              <SubPage pageContent={content} />
+              <SubPage />
             </Route>
             <Route path="/">
               <Helmet>
@@ -78,7 +38,7 @@ export default () => {
                   <img src={`${resolveAssetUrl("42UqcWCsSZ4IZyvq0Ucy6b")}?w=300`} alt="Turun Wappuradio" />
                 </div>
               }
-              <IndexPage content={indexContent} />
+              <IndexPage />
               {/* {process.env.REACT_APP_BROADCAST_MODE === 'live' && <RadioPlayer />}
               {process.env.REACT_APP_BROADCAST_MODE === 'live' && <VideoChatHider />}
               <ShowList shows={showList} /> */}
@@ -87,7 +47,7 @@ export default () => {
         </div>
       </Router>
       {/* <Sponsors /> */}
-      <Footer className="Footer" />
+      <Footer />
     </div>
   );
 }
